+++ wazo_calld/plugins/calls/plugin.py
@@ -5,6 +5,7 @@ from wazo_auth_client import Client as AuthClient
 from wazo_confd_client import Client as ConfdClient
 from wazo_amid_client import Client as AmidClient
 from xivo.pubsub import CallbackCollector
+from wazo_calld.phoned import PhonedClient
 
 from .bus_consume import CallsBusEventHandler
 from .notifier import CallNotifier
@@ -14,6 +15,8 @@ from .http import (
     CallMuteStartResource,
     CallMuteStopResource,
     CallDtmfResource,
+    CallHoldResource,
+    CallUnholdResource,
     CallsResource,
     ConnectCallToUserResource,
     MyCallResource,
@@ -21,6 +24,8 @@ from .http import (
     MyCallMuteStartResource,
     MyCallMuteStopResource,
     MyCallDtmfResource,
+    MyCallHoldResource,
+    MyCallUnholdResource,
 )
 from .services import CallsService
 from .stasis import CallsStasis
@@ -42,13 +47,15 @@ class Plugin:
 
         auth_client = AuthClient(**config['auth'])
         confd_client = ConfdClient(**config['confd'])
+        phoned_client = PhonedClient(**config['phoned'])
 
         token_changed_subscribe(confd_client.set_token)
+        token_changed_subscribe(phoned_client.set_token)
 
         dial_echo_manager = DialEchoManager()
 
         notifier = CallNotifier(bus_publisher)
-        calls_service = CallsService(amid_client, config['ari']['connection'], ari.client, confd_client, dial_echo_manager, notifier)
+        calls_service = CallsService(amid_client, config['ari']['connection'], ari.client, confd_client, dial_echo_manager, phoned_client, notifier)
 
         calls_stasis = CallsStasis(ari, collectd, bus_publisher, calls_service, config['uuid'], amid_client)
 
@@ -74,8 +81,12 @@ class Plugin:
         api.add_resource(CallMuteStartResource, '/calls/<call_id>/mute/start', resource_class_args=[calls_service])
         api.add_resource(CallMuteStopResource, '/calls/<call_id>/mute/stop', resource_class_args=[calls_service])
         api.add_resource(CallDtmfResource, '/calls/<call_id>/dtmf', resource_class_args=[calls_service])
+        api.add_resource(CallHoldResource, '/calls/<call_id>/hold/start', resource_class_args=[calls_service])
+        api.add_resource(CallUnholdResource, '/calls/<call_id>/hold/stop', resource_class_args=[calls_service])
         api.add_resource(MyCallResource, '/users/me/calls/<call_id>', resource_class_args=[auth_client, calls_service])
         api.add_resource(MyCallMuteStartResource, '/users/me/calls/<call_id>/mute/start', resource_class_args=[auth_client, calls_service])
         api.add_resource(MyCallMuteStopResource, '/users/me/calls/<call_id>/mute/stop', resource_class_args=[auth_client, calls_service])
         api.add_resource(MyCallDtmfResource, '/users/me/calls/<call_id>/dtmf', resource_class_args=[auth_client, calls_service])
+        api.add_resource(MyCallHoldResource, '/users/me/calls/<call_id>/hold/start', resource_class_args=[auth_client, calls_service])
+        api.add_resource(MyCallUnholdResource, '/users/me/calls/<call_id>/hold/stop', resource_class_args=[auth_client, calls_service])
         api.add_resource(ConnectCallToUserResource, '/calls/<call_id>/user/<user_uuid>', resource_class_args=[calls_service])