+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 20.11
+
+* New API to hold calls
+
+  * `PUT /calls/{call_id}/hold/start`
+  * `PUT /calls/{call_id}/hold/stop`
+  * `PUT /users/me/calls/{call_id}/hold/start`
+  * `PUT /users/me/calls/{call_id}/hold/stop`
+
 ## 20.08
 
 * Deprecate SSL configuration